Fix GH API path to include GH user
The API call seems to miss the GH user and returns always `null` for me when running in terminal. 

Note: I have no idea what that `:-` does in line 34 + 35, can you enlighten me? 

For the API call on GH it should use the original GH user + repo name otherwise the to lowercased values (for DockerHub) will fail in this case? 

Maybe this fixes #2 also. I had `description: true` and actually a description on my repo but it failed.